refactor: rename FlowCanvas and RunViewFlowCanvas
