+++ frontend/src/components/canvas/EditorCanvas.tsx
