+++ frontend/src/components/canvas/TraceCanvas.tsx
