+++ frontend/src/hooks/useKeyboardShortcuts.ts
@@ -1,6 +1,6 @@
 import { useEffect, useCallback, useState } from 'react'
 import { v4 as uuidv4 } from 'uuid'
-import { addNode, setNodes } from '../store/flowSlice'
+import { addNode, setNodes, setSelectedNode } from '../store/flowSlice'
 import { FlowWorkflowNode } from '@/types/api_types/nodeTypeSchemas'
 import { createNode } from '../utils/nodeFactory'
 import { AppDispatch } from '../store/store' // Import AppDispatch type
@@ -19,6 +19,18 @@ export const useKeyboardShortcuts = (
 
     const handleKeyDown = useCallback(
         (event: KeyboardEvent) => {
+            // Check if target is an input or textarea to avoid interfering with typing
+            const target = event.target as HTMLElement
+            const tagName = target.tagName.toLowerCase()
+            const isEditing = target.isContentEditable || tagName === 'input' || tagName === 'textarea' || tagName === 'select'
+
+            // Handle Escape key to close the sidebar if a node is selected
+            if (event.key === 'Escape' && selectedNodeID && !isEditing) {
+                event.preventDefault()
+                dispatch(setSelectedNode({ nodeId: null }))
+                return
+            }
+
             if (event.metaKey || event.ctrlKey) {
                 switch (event.key) {
                     case 'c': // CMD + C or CTRL + C
